# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Oscilloscope Backend API
  version: 1.0.0
  description: |
    Cloud API and documentation surface for the Hantek Oscilloscope Desktop application.

    ## Architecture

    ```
    Vercel Cloud  (this service)
      ✓  Swagger / OpenAPI interactive documentation
      ✓  Health check   GET /health
      ✓  Simulated signal demo   GET /signal/batch
      ✓  RPC protocol reference  GET /rpc-reference
      ✗  No real hardware access

    Windows Local Machine  (Electron Desktop App)
      ✓  Hantek DLL  (HTHardDll.dll / HT6000.dll)  via USB
      ✓  Python hardware_service.py  (JSON-RPC over stdio)
      ✓  Real-time ADC capture  up to 1 MSa/s
      ✓  Waveform + FFT display
      ✓  12-point hardware verification report
    ```

    Hardware-only endpoints return:
    > "Hardware operations require the local Electron Windows app."

  contact:
    name: Oscilloscope Desktop
  license:
    name: MIT
servers:
  - url: https://oscilloscope-backend.vercel.app
    description: Vercel cloud deployment
  - url: http://127.0.0.1:8765
    description: Local development server (uvicorn)
...
paths:
...
  /dll/select:
    post:
      tags: [dll]
      summary: Record DLL path
      description: |
        Records the path to the Hantek SDK DLL for this session.

        In the Electron app this triggers a file-system check and updates the
        hardware status panel.  In cloud mode the path is stored in memory only —
        no file-system operations are performed.
      operationId: dll_select
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectDllRequest'
            example:
              dll_path: "C:/Hantek/HTHardDll.dll"
      responses:
        "200":
          description: DLL path recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcResponse'
        "422":
          $ref: '#/components/responses/ValidationError'
...
  /device/connect:
    post:
      tags: [device]
      summary: Connect hardware device (Hardware Only)
      description: |
        **Hardware Only.**  Loads the Hantek DLL and calls `HT_OpenDevice` with
        the specified device index and export map.

        On success the device handle is retained for subsequent capture calls.
        Requires the Electron app on a Windows machine with the Hantek oscilloscope
        connected via USB.
      operationId: device_connect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRequest'
      responses:
        "200":
          description: Connection result or hardware-only message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcResponse'
        "422":
          $ref: '#/components/responses/ValidationError'
...
  /capture/start:
    post:
      tags: [capture]
      summary: Start real-time capture (Hardware Only)
      description: |
        **Hardware Only.**  Starts the background capture thread that continuously
        calls `HT_ReadData` and emits `sample_batch` events to the Electron renderer.

        `chunk_size` controls how many ADC samples are read per DLL call (default 2048).
        Requires an open device connection.
      operationId: capture_start
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartCaptureRequest'
      responses:
        "200":
          description: Capture start result or hardware-only message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcResponse'
        "422":
          $ref: '#/components/responses/ValidationError'
...
components:
  schemas:
    SelectDllRequest:
      type: object
      required: [dll_path]
      properties:
        dll_path:
          type: string
          description: Absolute or relative path to the Hantek SDK DLL
          example: "C:/Hantek/HTHardDll.dll"
    MapFunctionsRequest:
      type: object
      description: Maps logical SDK function names to actual DLL export names
      properties:
        initialize:
          type: string
          default: HT_Init
          description: DLL export name for SDK initialization
        open_device:
          type: string
          default: HT_OpenDevice
          description: DLL export name for opening the device
        close_device:
          type: string
          default: HT_CloseDevice
          description: DLL export name for closing the device
        start_capture:
          type: string
          default: HT_StartCapture
          description: DLL export name for starting ADC capture
        stop_capture:
          type: string
          default: HT_StopCapture
          description: DLL export name for stopping capture
        read_data:
          type: string
          default: HT_ReadData
          description: DLL export name for reading ADC sample data
    ConnectRequest:
      type: object
      properties:
        device_index:
          type: integer
          default: 0
          minimum: 0
          description: USB device index (0 = first connected oscilloscope)
        export_map:
          $ref: '#/components/schemas/MapFunctionsRequest'
    StartCaptureRequest:
      type: object
      properties:
        chunk_size:
          type: integer
          default: 2048
          minimum: 64
          maximum: 65536
          description: ADC samples per DLL read call
    HardwareStatus:
      type: object
      description: Complete hardware verification state
      properties:
        dll_path:
          type: string
          description: Recorded DLL file path
        dll_found:
          type: boolean
          description: DLL file exists on disk
        dll_arch:
          type: string
          enum: [x64, x86, unknown]
          description: DLL architecture
        python_arch:
          type: string
          description: Python process architecture
        electron_arch:
          type: string
          description: Electron process architecture (x64/ia32)
        dll_loaded:
          type: boolean
          description: DLL successfully loaded into process memory
        exports:
          type: array
          items:
            type: string
          nullable: true
          description: Exported function names discovered in DLL
        mapped_functions:
          type: object
          additionalProperties:
            type: string
          nullable: true
          description: Logical name → DLL export name mapping
        device_detected:
          type: boolean
          description: Hantek device detected on USB bus
        device_connected:
          type: boolean
          description: Device is open and handle is valid
        capture_started:
          type: boolean
          description: Capture started without SDK error
        capture_running:
          type: boolean
          description: Capture thread is alive and reading samples
        real_data_received:
          type: boolean
          description: At least one non-zero ADC sample batch received
        zero_data_warning:
          type: boolean
          description: ADC output consistently zero (cable / probe issue?)
        samples_received:
          type: integer
          description: Total ADC samples received since capture started
        last_min:
          type: integer
          nullable: true
          description: Minimum ADC code in last batch
        last_max:
          type: integer
          nullable: true
          description: Maximum ADC code in last batch
        last_variance:
          type: number
          format: double
          nullable: true
          description: Variance of last sample batch
        sdk_error_code:
          type: integer
          nullable: true
          description: Last Hantek SDK error code (0 = SUCCESS)
        last_error:
          type: string
          nullable: true
          description: Last error message from SDK or service
        final_status:
          type: string
          example: REAL HARDWARE VERIFIED
          description: 12-point verification result
        verification_reason:
          type: string
          description: Human-readable explanation of final_status
    RpcResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          description: Whether the operation succeeded
        message:
          type: string
          nullable: true
          description: Human-readable message
        status:
          allOf:
            - $ref: '#/components/schemas/HardwareStatus'
          nullable: true
          description: Current hardware status snapshot
...
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
  responses:
    ValidationError:
      description: Request body validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

# ── JSON-RPC Protocol Reference (informational / non-HTTP) ────────────────────
#
# Transport : Electron spawns Python via child_process.spawn
# Request   : stdin  newline-delimited JSON  { id, method, params }
# Response  : stdout newline-delimited JSON  { event: "rpc_response", payload: { request_id, ok, result, error } }
# Events    : stdout newline-delimited JSON  { event: "status"|"sample_batch"|"log", payload: {...} }
#
# Methods:
#   select_dll        { dll_path }                  → status event
#   inspect_dll       {}                             → status event  (Windows only)
#   map_sdk_functions { initialize, open_device, close_device, start_capture, stop_capture, read_data }
#   connect_device    { device_index, export_map }  → status event  (Windows + DLL required)
#   start_capture     { chunk_size }                → status + sample_batch stream
#   stop_capture      {}                             → status event
#   get_status        {}                             → rpc_response with status
#   disconnect_device {}                             → status event
#   shutdown          {}                             → (process exits)
#
# sample_batch payload: { timestamp_unix, samples: [int16], count, min, max, variance }
# status payload      : HardwareStatus schema above
# log payload         : { level: "INFO|WARNING|ERROR|DEBUG", message: str }